backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.35

psycopg[binary]==3.2.9
alembic==1.13.2
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
pyyaml==6.0.2
python-multipart==0.0.9
boto3==1.35.10
pytesseract==0.3.13
pdf2image==1.17.0        # requires poppler in image (prod Dockerfile)
Pillow==10.4.0
qdrant-client==1.10.1
sentence-transformers==3.0.1
httpx==0.27.2
minio>=7.2.7
